requirements.txt
cffi==1.17.1
clr_loader==0.2.7.post0
pycparser==2.22
pythonnet==3.0.5
